Fixes issue #1.
If logfile doesn't exist[1], it is created and function is run again.
Technically a case of recursion, although not expected to occour more than once.

Also noticed failure should logfile not exist[1] and log summary be requested.
This issue is now bypassed, and log error will be communicated in advance.

[1] Not exist, or be badly formatted.